Unirest object mapper moved to shared utils so we an re-use it if need be later on. Cleaned some more other things to make it more readable and easier to understand (SOLID)
